chore: replace multiple reader files a single
