--- src/core/mantle/data/transitReader/index.ts
@@ -1,2 +1,0 @@
-export { GTFSReader } from "./gtfsReader";
-export { GTFS, Trips, Trip, OccupancyStatus } from "./gtfsInterface";
